requirements.txt
pandas
numpy
scikit-learn
streamlit
plotly
streamlit-autorefresh
python-dotenv
openai
requests
